requirements.txt
# Core
python-dotenv==1.0.0
loguru==0.7.2

# Módulo 1 - Ingestão de Dados
pypdf==3.17.4
PyMuPDF==1.23.26
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
chardet==5.2.0
langchain-text-splitters==0.0.1
openpyxl==3.1.2

# Módulo 2 - Benchmark
sentence-transformers==2.3.1
qdrant-client==1.7.3
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# Módulo 3 - RAG
ollama==0.1.6
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Development & Testing
pytest==7.4.4
pytest-benchmark==4.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

